Add a self-timer
Summary:
Adds a self timer with the options off, 2 s, 5 s 10 s and 20 s.
It can be used in photo and video mode.

When the self-timer is running, a countown and a flashing red border
are shown, the capture/recording action changes to a cancel action.

The functionality of the onTriggered handler of the captureAction has
been moved to the onTriggered handler of the selfTimer.

Reviewers: #plasma:_mobile, jbbgameich

Reviewed By: jbbgameich

Subscribers: jbbgameich

Differential Revision: https://phabricator.kde.org/D23234