Add software rotation based on what the camera reports
See https://libcamera.org/api-html/namespacelibcamera_1_1properties.html#a3b25103242b577663c87eb8d492fe94e

Libcamera reports the mounting rotation of the camera. We can use this
to rotate our frames so that they match up with the screen orientation.
We also need to account for monitor orientation as another layer on top
of this (not yet implemented in this MR). This is necessary for phones since their cameras are mounted
according to the long side of the phone.

Hardware rotation in libcamera appears to not be well supported by
drivers, my webcams don't seem to have it, and it crashes the app on the
OnePlus 6: https://libcamera.org/api-html/classlibcamera_1_1CameraConfiguration.html#a44acd9a3af5ce3a937c008e0843268b4